Transaction isolation. #272

When client transactions were implemented the DB's read-write lock was removed, exposing H2's transaction isolation semantics. The default isolation level is "read committed", which allows concurrent transactions to overwrite each other's updates. This was the cause of #272. Changing H2's isolation level to "serialisable" would have caused other problems.

The solution is to reintroduce the DB's read-write lock. The lock is acquired when starting a transaction and released when committing or rolling back a transaction. (We already use try/finally blocks to ensure every transaction is committed or rolled back.) Read-only transactions can share the lock. To avoid deadlock, transactions must not be started while holding other locks.

This patch adapts the key manager to the new locking rules. The rest of the code was already compliant. Transports are now added to the DB during the startup phase, which allows TransportAddedEvent and TransportRemovedEvent to be deleted.

Fixes #269, fixes #272.

See merge request !124

+++ briar-core/src/org/briarproject/db/DatabaseComponentImpl.java
@@ -33,8 +33,6 @@ import org.briarproject.api.event.MessageValidatedEvent;
 import org.briarproject.api.event.MessagesAckedEvent;
 import org.briarproject.api.event.MessagesSentEvent;
 import org.briarproject.api.event.SettingsUpdatedEvent;
-import org.briarproject.api.event.TransportAddedEvent;
-import org.briarproject.api.event.TransportRemovedEvent;
 import org.briarproject.api.identity.Author;
 import org.briarproject.api.identity.AuthorId;
 import org.briarproject.api.identity.LocalAuthor;
@@ -61,6 +59,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.locks.ReadWriteLock;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.logging.Logger;
 
 import javax.inject.Inject;
@@ -80,6 +80,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 	private final EventBus eventBus;
 	private final ShutdownManager shutdown;
 	private final AtomicBoolean closed = new AtomicBoolean(false);
+	private final ReadWriteLock lock = new ReentrantReadWriteLock(true);
 
 	private volatile int shutdownHandle = -1;
 
@@ -117,18 +118,33 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 		db.close();
 	}
 
-	public Transaction startTransaction() throws DbException {
-		return new Transaction(db.startTransaction());
+	public Transaction startTransaction(boolean readOnly) throws DbException {
+		if (readOnly) lock.readLock().lock();
+		else lock.writeLock().lock();
+		try {
+			return new Transaction(db.startTransaction(), readOnly);
+		} catch (DbException e) {
+			if (readOnly) lock.readLock().unlock();
+			else lock.writeLock().unlock();
+			throw e;
+		} catch (RuntimeException e) {
+			if (readOnly) lock.readLock().unlock();
+			else lock.writeLock().unlock();
+			throw e;
+		}
 	}
 
 	public void endTransaction(Transaction transaction) throws DbException {
-		T txn = txnClass.cast(transaction.unbox());
-		if (transaction.isComplete()) {
-			db.commitTransaction(txn);
-			for (Event e : transaction.getEvents()) eventBus.broadcast(e);
-		} else {
-			db.abortTransaction(txn);
+		try {
+			T txn = txnClass.cast(transaction.unbox());
+			if (transaction.isComplete()) db.commitTransaction(txn);
+			else db.abortTransaction(txn);
+		} finally {
+			if (transaction.isReadOnly()) lock.readLock().unlock();
+			else lock.writeLock().unlock();
 		}
+		if (transaction.isComplete())
+			for (Event e : transaction.getEvents()) eventBus.broadcast(e);
 	}
 
 	private T unbox(Transaction transaction) {
@@ -138,6 +154,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public ContactId addContact(Transaction transaction, Author remote,
 			AuthorId local, boolean active) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsLocalAuthor(txn, local))
 			throw new NoSuchLocalAuthorException();
@@ -150,6 +167,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 	}
 
 	public void addGroup(Transaction transaction, Group g) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsGroup(txn, g.getId())) {
 			db.addGroup(txn, g);
@@ -159,6 +177,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void addLocalAuthor(Transaction transaction, LocalAuthor a)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsLocalAuthor(txn, a.getId())) {
 			db.addLocalAuthor(txn, a);
@@ -168,6 +187,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void addLocalMessage(Transaction transaction, Message m, ClientId c,
 			Metadata meta, boolean shared) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsGroup(txn, m.getGroupId()))
 			throw new NoSuchGroupException();
@@ -191,15 +211,15 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void addTransport(Transaction transaction, TransportId t,
 			int maxLatency) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
-		if (!db.containsTransport(txn, t)) {
+		if (!db.containsTransport(txn, t))
 			db.addTransport(txn, t, maxLatency);
-			transaction.attach(new TransportAddedEvent(t, maxLatency));
-		}
 	}
 
 	public void addTransportKeys(Transaction transaction, ContactId c,
 			TransportKeys k) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -210,6 +230,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void deleteMessage(Transaction transaction, MessageId m)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsMessage(txn, m))
 			throw new NoSuchMessageException();
@@ -218,6 +239,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void deleteMessageMetadata(Transaction transaction, MessageId m)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsMessage(txn, m))
 			throw new NoSuchMessageException();
@@ -226,6 +248,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public Ack generateAck(Transaction transaction, ContactId c,
 			int maxMessages) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -237,6 +260,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public Collection<byte[]> generateBatch(Transaction transaction,
 			ContactId c, int maxLength, int maxLatency) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -254,6 +278,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public Offer generateOffer(Transaction transaction, ContactId c,
 			int maxMessages, int maxLatency) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -265,6 +290,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public Request generateRequest(Transaction transaction, ContactId c,
 			int maxMessages) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -277,6 +303,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public Collection<byte[]> generateRequestedBatch(Transaction transaction,
 			ContactId c, int maxLength, int maxLatency) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -420,14 +447,9 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 		return db.getTransportKeys(txn, t);
 	}
 
-	public Map<TransportId, Integer> getTransportLatencies(
-			Transaction transaction) throws DbException {
-		T txn = unbox(transaction);
-		return db.getTransportLatencies(txn);
-	}
-
 	public void incrementStreamCounter(Transaction transaction, ContactId c,
 			TransportId t, long rotationPeriod) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -448,6 +470,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void mergeGroupMetadata(Transaction transaction, GroupId g,
 			Metadata meta) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsGroup(txn, g))
 			throw new NoSuchGroupException();
@@ -456,6 +479,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void mergeMessageMetadata(Transaction transaction, MessageId m,
 			Metadata meta) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsMessage(txn, m))
 			throw new NoSuchMessageException();
@@ -464,6 +488,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void mergeSettings(Transaction transaction, Settings s,
 			String namespace) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		Settings old = db.getSettings(txn, namespace);
 		Settings merged = new Settings();
@@ -477,6 +502,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void receiveAck(Transaction transaction, ContactId c, Ack a)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -492,6 +518,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void receiveMessage(Transaction transaction, ContactId c, Message m)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -507,6 +534,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void receiveOffer(Transaction transaction, ContactId c, Offer o)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -529,6 +557,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void receiveRequest(Transaction transaction, ContactId c, Request r)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -545,6 +574,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void removeContact(Transaction transaction, ContactId c)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -554,6 +584,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void removeGroup(Transaction transaction, Group g)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		GroupId id = g.getId();
 		if (!db.containsGroup(txn, id))
@@ -566,6 +597,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void removeLocalAuthor(Transaction transaction, AuthorId a)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsLocalAuthor(txn, a))
 			throw new NoSuchLocalAuthorException();
@@ -575,15 +607,16 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void removeTransport(Transaction transaction, TransportId t)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsTransport(txn, t))
 			throw new NoSuchTransportException();
 		db.removeTransport(txn, t);
-		transaction.attach(new TransportRemovedEvent(t));
 	}
 
 	public void setContactActive(Transaction transaction, ContactId c,
 			boolean active) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -593,6 +626,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void setMessageShared(Transaction transaction, Message m,
 			boolean shared) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsMessage(txn, m.getId()))
 			throw new NoSuchMessageException();
@@ -602,6 +636,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void setMessageValid(Transaction transaction, Message m, ClientId c,
 			boolean valid) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsMessage(txn, m.getId()))
 			throw new NoSuchMessageException();
@@ -612,6 +647,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 	public void setReorderingWindow(Transaction transaction, ContactId c,
 			TransportId t, long rotationPeriod, long base, byte[] bitmap)
 			throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -622,6 +658,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void setVisibleToContact(Transaction transaction, ContactId c,
 			GroupId g, boolean visible) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		if (!db.containsContact(txn, c))
 			throw new NoSuchContactException();
@@ -647,6 +684,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
 	public void updateTransportKeys(Transaction transaction,
 			Map<ContactId, TransportKeys> keys) throws DbException {
+		if (transaction.isReadOnly()) throw new IllegalArgumentException();
 		T txn = unbox(transaction);
 		Map<ContactId, TransportKeys> filtered =
 				new HashMap<ContactId, TransportKeys>();
